Final App
after trying with many other plugins decided to go with a google calendar implementation. that worked on all devices, but then i saw a cordova plugin that could work natively on android and IOS, it did, but then the web app was not functional.
Since i had already the google calendar implementation working, i used an interface to switch between Google and Native Calendar. the architecture will be available in the readme file